LayoutTests/html5lib/runner-expected-html5.txt

CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 1: SyntaxError: Parse error
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 0: Warning: Problem parsing viewBox=""
CONSOLE MESSAGE: line 2: PASS
CONSOLE MESSAGE: line 2: FOO<span>BAR</span>BAZ
Content-Type: text/plain
resources/tests1.dat: PASS

resources/tests2.dat: PASS

resources/tests3.dat: PASS

resources/tests4.dat:
3
5
6
7

Test 3 of 7 in resources/tests4.dat failed. Input:
textarea content with <em>pseudo</em> <foo>markup
Got:
| "textarea content with "
| <em>
|   "pseudo"
| " "
| <foo>
|   "markup"
Expected:
| "textarea content with <em>pseudo</em> <foo>markup"

Test 5 of 7 in resources/tests4.dat failed. Input:
</plaintext>
Got:
| 
Expected:
| "</plaintext>"

Test 6 of 7 in resources/tests4.dat failed. Input:
setting html's innerHTML
Got:
| "setting html's innerHTML"
Expected:
| <head>
| <body>
|   "setting html's innerHTML"

Test 7 of 7 in resources/tests4.dat failed. Input:
<title>setting head's innerHTML</title>
Got:
| 
Expected:
| <title>
|   "setting head's innerHTML"
resources/tests5.dat: PASS

resources/tests6.dat:
27
30
37
44
45
47

Test 27 of 51 in resources/tests6.dat failed. Input:
foo<col>
Got:
| 
Expected:
| <col>

Test 30 of 51 in resources/tests6.dat failed. Input:
</frameset><frame>
Got:
| 
Expected:
| <frame>

Test 37 of 51 in resources/tests6.dat failed. Input:
<caption><col><colgroup><tbody><tfoot><thead><tr>
Got:
| <caption>
| <colgroup>
| <tbody>
| <tfoot>
| <thead>
|   <tr>
Expected:
| <tr>

Test 44 of 51 in resources/tests6.dat failed. Input:
</table><tr>
Got:
| <tr>
Expected:
| <tbody>
|   <tr>

Test 45 of 51 in resources/tests6.dat failed. Input:
<body></body></html>
Got:
| 
Expected:
| <head>
| <body>

Test 47 of 51 in resources/tests6.dat failed. Input:
<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01//EN"><html></html>
Got:
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
Expected:
| <!DOCTYPE html "-//W3C//DTD HTML 4.01//EN" "">
| <html>
|   <head>
|   <body>
resources/tests7.dat:
24
30

Test 24 of 30 in resources/tests7.dat failed. Input:
<body>X</body></body>
Got:
| "X"
Expected:
| <head>
| <body>
|   "X"

Test 30 of 30 in resources/tests7.dat failed. Input:
<select><keygen>
Got:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
|       <option>
|         "2048 (High Grade)"
|       <option>
|         "1024 (Medium Grade)"
|       <option>
|         "512 (Low Grade)"
Expected:
| <html>
|   <head>
|   <body>
|     <select>
|     <keygen>
resources/tests8.dat: PASS

resources/tests9.dat: PASS

resources/tests10.dat: PASS

resources/tests11.dat: PASS

resources/tests12.dat: PASS

resources/tests14.dat: PASS

resources/tests15.dat: PASS

resources/tests16.dat: PASS

resources/webkit01.dat: PASS

resources/doctype01.dat: PASS

resources/scriptdata01.dat: PASS

resources/entities01.dat:
2
5

Test 2 of 68 in resources/entities01.dat failed. Input:
FOO&gtBAR
Got:
| <html>
|   <head>
|   <body>
|     "FOO&gtBAR"
Expected:
| <html>
|   <head>
|   <body>
|     "FOO>BAR"

Test 5 of 68 in resources/entities01.dat failed. Input:
I'm &notit; I tell you
Got:
| <html>
|   <head>
|   <body>
|     "I'm &notit; I tell you"
Expected:
| <html>
|   <head>
|   <body>
|     "I'm ¬it; I tell you"
resources/entities02.dat: PASS

resources/comments01.dat: PASS

resources/adoption01.dat:
3
11
13

Test 3 of 13 in resources/adoption01.dat failed. Input:
<a>1<button>2</a>3</button>
Got:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "23"
Expected:
| <html>
|   <head>
|   <body>
|     <a>
|       "1"
|       <button>
|         "2"
|     "3"

Test 11 of 13 in resources/adoption01.dat failed. Input:
<p><b id="A"><script>document.getElementById("A").id = "B"</script></p>TEXT</b>
Got:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="B"
|       "TEXT"
Expected:
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         id="B"
|         <script>
|           "document.getElementById("A").id = "B""
|     <b>
|       id="A"
|       "TEXT"

Test 13 of 13 in resources/adoption01.dat failed. Input:
<table>A<td>B</td>C</table>
Got:
| <html>
|   <head>
|   <body>
|     "AC"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
Expected:
| <html>
|   <head>
|   <body>
|     "A"
|     "C"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "B"
resources/inbody01.dat: PASS

resources/isindex.dat: PASS

resources/tables01.dat: PASS
#EOF
